+++ ModConfig.cs
@@ -210,8 +210,10 @@ sealed class ModConfig(ConfigFile cfg)
     {
         public ConfigEntry<bool> Enable { get; } = cfg.Bind(section, nameof(Enable), false, "True to make dropped trophies spawn mobs");
         public ConfigEntry<int> ActivationDelay { get; } = cfg.Bind(section, nameof(ActivationDelay), 3660, "Time in seconds before trophies start spawning mobs.");
-        public ConfigEntry<int> MinRespawnDelay { get; } = cfg.Bind(section, nameof(MinRespawnDelay), 12, "Minimum respawn delay in seconds");
-        public ConfigEntry<int> MaxRespawnDelay { get; } = cfg.Bind(section, nameof(MaxRespawnDelay), 240, "Maximum respawn delay in seconds");
+        public ConfigEntry<int> MinRespawnDelay { get; } = cfg.Bind(section, nameof(MinRespawnDelay), 12,
+            "Minimum respawn delay in seconds. The actual respawn delay is chosen between min and max based on the stack size of the dropped trophies");
+        public ConfigEntry<int> MaxRespawnDelay { get; } = cfg.Bind(section, nameof(MaxRespawnDelay), 240,
+            "Maximum respawn delay in seconds. The actual respawn delay is chosen between min and max based on the stack size of the dropped trophies");
     }
 
     public sealed class WorldModifiersConfig(ConfigFile cfg, string section)